Solution to Check MSB of a Number

+++ Check MSB of a Number@@67882a731b4392d8fc1577e9/code.c
@@ -5,7 +5,7 @@ char* welcome() {
 }
 
 int main() {
-    float a;
+    long a;
     scanf("%ld",&a);
     (a&4294967296 == 4294967296)?printf("Set"):printf("Not Set");
     return 0;
